+++ daniel_montecinos/src/main/resources/phase-1.txt
