--- daniel_montecinos/tic_tac_toe/.gitignore
@@ -1,10 +1,0 @@
-.gradle/
-.idea/
-build
-out/
-config/
-gradle/
-build.gradle
-gradlew.bat
-gradlew
-*.iml
